refactor: ladder클래스에 존재하는 drawName 메서드 players클래스로 이동
