+++ src/main/java/kr/codesquad/domain/Players.java
@@ -54,4 +54,18 @@ public class Players {
             throw new RuntimeException(PLAYER_NUMBER_ERROR_MESSAGE);
         }
     }
+
+    private void drawName(StringBuilder sb) {
+        for (String name : nameList) {
+            sb.append(String.format(" %1$-5s", name));
+        }
+        sb.append("\n");
+    }
+
+    @Override
+    public String toString() {
+        StringBuilder sb = new StringBuilder();
+        drawName(sb);
+        return sb.toString();
+    }
 }
