+++ src/main/java/kr/codesquad/view/OutputView.java
@@ -1,10 +1,11 @@
 package kr.codesquad.view;
 
 import kr.codesquad.domain.Ladder;
+import kr.codesquad.domain.Players;
 import kr.codesquad.domain.Result;
 
 public class OutputView {
-    public void printLadder(Ladder ladder, Result result) {
-        System.out.println(ladder+"\n"+ result);
+    public void printLadder(Ladder ladder, Players player,Result result) {
+        System.out.println(player+""+ladder+""+ result);
     }
 }
